Testing and docstring tidy up
